requirements.txt
vosk==0.3.44
sounddevice==0.4.7
paho-mqtt==2.1.0
numpy
